requirements.txt
Flask>=3.0
Flask-Login>=0.6
Werkzeug>=3.0
playwright>=1.40
pandas>=2.0
openpyxl>=3.1
beautifulsoup4>=4.12
requests>=2.31
python-dotenv>=1.0
